feat(consensus): Download consensus file and move tests
